GitMoveDeleteHook: Cancel move if it would affect unmerged file
Before this change, the index entries belonging to the unmerged path
would get into a "funny" state.

The only good alternative to canceling the move would be to preserve the
index entries under the new path. C Git aborts with an error when doing
"git mv", so we do the same now.

Bug: 391857
Change-Id: I0477f62c9f4997890ea15f1228413f33473577c8