Recognize a gerrit fetch command in the Fetch From Gerrit dialog
Gerrit can populate the clipboard with a command to fetch a change from
Gerrit using the command line. One of these, i.e. the checkout
FETCH_HEAD, matches the capabilities of the fetch dialog. If that is
the command populate all fields accordingly so the user can just press
"Finish".

Change-Id: Id2e6f1eeb8c6d0bf89564ea78f6f71f9e9257d49