Improve precision of detached head decoration
For decorations in EGit we want to display the name used for checking
out a tag. If a tag was checked out HEAD points to a commitId directly,
but the reflog contains the tag name used. Scan the reflog and use the
latest checkout whose refname resolves to the same commit as HEAD.

If not try, as before, to find any matching ref and use the one last
updated.

Bug: 385392
Change-Id: Icb10545a1e503a98fd9ea98fbb6770c37a9e4bfe
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>