Eliminate file handle leaks
Running the SWTBot test suite on a Windows environment highlights
insidious file handle leaks caused by JGit's WindowCache which keeps
".pack" files open for reading and thus locked for deletion until the
repository is properly closed, but EGit might not keep a reference
towards this repository in its cache, "forgetting" to close it when
needed. This is mostly true for sub-modules.

This patch also gets rid of other random errors, sometimes due to a
.project not properly removed.

Bug: 404121
Change-Id: I3a596fcabdb5225e9042e1ffae6b054c48d72d54
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>