[stagingView] Fix spurious warning if view was hidden when input set
See bug for how to reproduce. The fix consists of two parts:

1. Change isVisible to getVisible.

For isVisible to be true, the parents of also have to be visible. In the
case of the warning label, we also want to correctly toggle it even if
e.g. the staging view is not currently visible.

2. Rework modify handlers

Instead of requiring users of CommitMessageComponent to handle modify
events of author/committer texts themselves, handle it in the component.
The advantage of this is that the message does not have to be updated
when the listeners in the component are disabled. Before this, a warning
would theoretically be shown for a brief moment when authorText.setText
was called inside updateUI but the committer was not yet set.

The first part alone would have been enough to fix the problem. The
second part improves the code by making it simpler.

Bug: 413286
Change-Id: Ic46f1a5b0627fdca71cc0a83c28c3121ce519685
Signed-off-by: Robin Stocker <robin@nibor.org>