RepositoryAction: Only calculate isEnabled if relevant selection changes
When selecting text in a text editor or moving the cursor,
selectionChanged is called many times with an ITextSelection.

In case of a text selection, we are only interested in which file the
selection is in. To get rid of some unnecessary cycles, only calculate
isEnabled if the selection changed after the interesting parts are
extracted.

Bug: 415958
Change-Id: I52416d0386a60d15b19d0a32755b016c2f3419eb
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>