[historyView] Prevent nested rebase
Disable rebase when repository is not in a safe state. This prevents
calling rebase when there is already a rebase, merge or cherry-pick in
progress.

Also safeguard against broken handlers which may miss to become disabled
when not in a safe repository state in order to prevent that nested
rebase gets aborted since this could destroy the work done while
resolving conflicts in the first rebase. Throw an IllegalStateException
in this case.

Bug: 416136
Change-Id: I05054b6c7f6a49f4cd01d70ee44791e6985e9ee5
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>