[repoView] Improve performance of tags (label provider)
Before, parsing the tags was done in different places. For the image
(annotated vs lightweight), the tag was always parsed. For the check
mark when tag is checked out, the tag was parsed again. And in case of
the option "Display Latest Branch Commit" being enabled, the tag was
parsed a third time.

With this change, each tag is parsed only once in the content provider
and the result is stored in the TagNode. Also, it uses one RevWalk
instance instead of N.

Bug: 411759
Change-Id: Ic65db5b29a1a2897ec4f613692794365e4e9439b
Signed-off-by: Robin Stocker <robin@nibor.org>