Add a helper method to find repository from given resource path
In most cases we can use RepositoryMapping to find the repository
responsible for versioning a given resource. But the team framework
doesn't allow to use RepositoryMappings to find the repository for
resources in a closed project. Hence add a helper method to find the
repository based on comparison of resource path and the working tree
path of all repositories known to the RepositoryCache.

Change-Id: I8d26d5a6ed4aa6661eb88d540ab643fb3b6cebbc
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>