[stagingView] Fix rebase buttons being shown after startup
When the staging view is shown after startup and it's still
initializing, the rebase buttons were shown for a time, and then hidden
after initialization was finished.

The reason was that updateRebaseButtonVisibility schedules the hiding
using Display#asyncExec. When called from createPartControl, this is not
necessary, so hide it directly there.

Change-Id: Ibb6fc5f0fb2c28ace76159a4d71ad273718387c9
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>