Import: Connect projects from submodules to correct repo
Before, all projects were connected to the repository that was
cloned/selected. In case of projects that are inside submodules, this
would connect them to the super repository.

With this change, the repository is searched again (in parents) to find
the right one. If they are not within a submodule, the finder will
automatically find the original repo.

Bug: 423415
Change-Id: Ib4af9a6bf79b329ce59aa61002851a9568554551
Signed-off-by: Robin Stocker <robin@nibor.org>