Also show Gerrit actions if notes refspec is existing
So far Gerrit actions are enabled for a repository if the option
gerrit.createchangeid = true or a push refspec exists which has a
destination starting with 'refs/for/'.

Enhance this to also enable Gerrit actions if a fetch refspec exists
with both source and destination are starting with 'refs/notes/'.

Bug: 433439
Change-Id: Ifdc0aa044c1694bcdb808e205592fd2b40dba624
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>