Fix missing trailing newline with "Create Patch"
The resulting patches didn't work with "git apply" because the trailing
newline was missing. Adding it makes it work, and it also still works
with Eclipse's "Apply Patch".

Just removing the line with "trim newline" at the end wasn't enough,
because the ByteArrayOutputStream was only overriding one of the write
methods (not the one with just one char). So the last newline was never
added to the StringBuilder. Replace it with a more straightforward
implementation.

Bug: 406134
Change-Id: If736ab3bc5972908a464212a29de8311b32e2718
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>