Scalability: label decorator should cache computed repo branch state
In workspaces with lot of projects from the same repository the most
decorator time is spent on computing branch state of the same
repository. A repository/branch state weak hash map is introduced in
DecoratableResourceHelper to avoid redundant calculation of the
repository branch state. The map entries are invalidated by
GitLightweightDecorator on each index change.

Bug: 457431
Change-Id: I1de94280f2d3d45ae8ef28017dc08a9deadf3075
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>