Introduce compare mode for Git Staging view similar to History view
A compare mode was introduced. If it's switched off a double click on a
file in the staging view will open the file's workspace version in an
editor. Otherwise the double click opens a compare editor. The compare
mode can be configured using a new action in the view's toolbar or using
the view menu.

Furthermore an "Compare with Index" action was added to the pop-up menu
to allow the user to compare a file with the index.

Bug: 473486
Change-Id: Ifba134d4585d15dddebe16eb87ae9c2231e7cf97
Signed-off-by: Jan-Ove Weichel <ovi.weichel@gmail.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>