Fix labeling when Repository View is not open
The ResourceStateFactory must not rely exclusively on the ResourceCache
to find repositories. That cache is not guaranteed to contain all
configured repositories at all times. Only while the repository view is
open we can rely on the cache to contain all configured repositories.

Instead, use RepositoryMapping.getMapping, and fall back to the cache
only when that doesn't find anything.

This is a general problem with this cache. The RepositoryCache is
populated by three code paths:

* RepositoryMapping.getMapping() -- on-demand-load; no hard reference
  is retained permanently in general.
* Auto-sharing in the core Activator -- on-demand-load; no hard
  reference is retained permanently.
* The content provider of the Repositories View -- full load; hard
  reference kept as long as the view is open.

The fix here is minimal. A more complete fix would be to make
RepositoryCache.getRepository(IPath) demand-load missing configured
repositories. I _do_ have an implementation for that, too, but that
needs more testing.

Bug: 484110
Change-Id: I286e070955cc78193f6d9481c539939e7507b37d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>