Decorate IFolders that are submodule working directory roots
Introduce a new configurable submodule text decoration used for folders
that are submodule working directory roots. (Or working directory roots
of nested repositories.)

The default shows the a dirty indicator, the branch, branch state, and
the head commit's short message. The repository name is omitted since
it'll always be identical to the folder name.

Added a new variable for the short message of the head commit.

Change-Id: I75460fd59f2e6c9bd06e82a1966e3d06f97ab709
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>