Improve error reporting in FetchGerritChangePage
If a connection failure occurs during the change number content assist,
the user feedback was poor. Some quick flickering of the progress bar
and that was it. The error log contained a stack trace starting with two
InvocationTargetExceptions without exception messages, which isn't
exactly helpful either.

Improve this in two ways:

* Generally skip InvocationTargetExceptions with empty messages when
  reporting or logging errors.
* In the case of the content proposal in the FetchGerritChangePage,
  show the error to the user.
* Move the error reporting methods in Activator together. I completely
  missed at first that there was a createErrorStatus() further below...

Change-Id: I9fa6bb6207539cff1c98850f7abf5d127077115c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>