Eliminate defaultHandler for "Rename Branch..."
With a default handler for this command, there are two quick access
entries: "Rename Branch" and "Rename Branch...". They are available at
all times, but the first one should have the ellipsis, and the second
one does not work (throws an IndexOutOfBoundsException) if the selection
is in an unexpected part.

* Change the command name for the org.eclipse.egit.ui.team.RenameBranch
  command to "Rename Branch..." so that it show correctly in quick
  access.
* Remove the defaultHandlers for the various renameBranch commands. Add
  explicit handlers with correct <activeWhen> expressions.
* Remove the unused org.eclipse.egit.ui.RepositoriesViewRenameBranch
  command from plugin.xml.
* Make the CommitGraphTable use the standard rename command, like the
  repositories view does.
* Delete unused class RenameBranchAction.
* Simplify AbstractHistoryCommandHandler a little bit.

As a result, there is only one "Rename Branch..." available in quick
access, and only if applicable. Moreover, the correct handler is invoked
if applicable. Context menus are unchanged in appearance and
functionality.

To fix bug 495064, one would have to do similar refactorings for all
other default handlers. (Unless, of course, for those that really make
sense as a default handler.)

Bug: 495064
Change-Id: Ic85726d97f5e73ea3feb52a8fab420408d78bfc7
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>