Use jobs instead of threads for searching in history
The main reason for doing this is that I wanted to get rid of the custom
progress meter. It's also a first step towards using a CommitSearchQuery
instead of yet another custom-written search logic, and a first step
towards getting a more compact search widget that perhaps can be even
moved into the toolbar.

Also make the text field a search field, and react to ENTER and to the
search icon being clicked. Add a new listener interface IFindListener
to disentangle the toolbar, the job, and the find results.

Bug: 481019
Bug: 496278
Change-Id: I69e15163451c0be8606d48bedb49b1fd11ec74bf
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>